Basic setup, create grid
In this commit, I set up the HTML doc with links to the javascript and
CSS files. Then, I set up a div container, with which I used a for loop
in javascript to create 256 div elements. From there, I styled the divs
to be a visible grid of squares using CSS properties, including layout
properties in flexbox style.