Update HTML - groundwork for UI
In this commit, I have added containers, inputs, and labels for
the different input controls I want within my UI, including a selector
for the background color of the grid, a custom pen color, a random pen
color, an eraser, a button to clear the board, and a slider input for
grid size. I also attempted to correct an error in the javascript, to no
avail. Will return to this in future commits.