Change container ID to grid-container
This will help me with specificity, since I have a few divs that are
functioning as containers, and having a non-specific container label on
a div will not help me in the future. I did not correct this in the
javascript yet, because I will be undergoing an overhaul of the
javascript in future commits.