Add function for grid lines, clear grid
In this commit, I have added two similar functions with different
purposes. The first, toggleGridLines(), is meant to help the checkbox in
the UI turn the grid lines on or off. The second, clearGrid(), is
meant to clear all changes to the squares within the grid. I will work
to ensure the success of both of these functions in future commits.