begin() function
In this commit, I have added the begin() function, which effectively
takes the groundwork I have laid and enables its functionality. This
function adds event listeners to different areas of the UI, calls the
functions to create and resize the grid and/or toggle the grid lines on
or off, and allows the user to change the brush mode.