Color functionality
In this commit, I added the color functionalities to the javascript to
finalize the back-end of my Etch A Sketch project. In response to the
other extra credit prompt from The Odin Project, I also added a button
and associated function to enable the pen to select a random color for
each square that it passes through, when this button is selected.