Basic styles, troubleshooting
I am still having issues with functionality, in that the browser is not
creating the grid like it should. However, in this commit, I did fix
some issues in the HTML, add basic styling in the CSS (which I will
revisit at a later date), and correct a few noticed errors in the
javascript.